Splat once again compiles on Windows.
Updated CMakeLists.txt so that we can compile on Windows as well as
compile and link in libogg and libvorbis. Custom CMakeLists were added
for those projects as well as sdl_mixer (rather than use the one I
previously added sdl_mixer's own source tree).

Minor changes to the .gitignore file since we now need to track the
contents of sdl_mixer and MSVC solution and project files.

Executables are now all dropped in bin/ because it's better organized
that way and becuase this ensures that the SDL .dll files end up in the
right place along with the executable. Updated the various shell and
make scripts to look in the right place for executables.

Due to a compiler problem on windows, a number of places where we had
vectors of structures have been replaced by vectors of pointers to
structures. As a result, a bunch of places went from using . to ->, and
a bunch of functions now take pointers instead of references.

Removed references to windows headers that were pulled in to use `Sleep`.
Instead, we SDL_delay.

Suppressed some warnings in visual studio when compiling mathfu.

Changed a bunch of ints to size_t, uint8_t or uint16_t types as
appropriate because Visual Studio is really stingy about implicit
typecasts.

Tested on Linux and Windows.
Bug: 17515158 17514182 17515752 17514180 17515256

+++ tests/audio_engine/audio_engine_test.cpp
@@ -45,9 +45,9 @@ class AudioEngineTests : public ::testing::Test {
  protected:
   virtual void SetUp() {
     // Make a bunch of sound defs with various priorities.
-    for (int i = 0; i < 6; ++i) {
+    for (uint16_t i = 0; i < 6; ++i) {
       flatbuffers::FlatBufferBuilder builder;
-      int id = i;
+      uint16_t id = i;
       float priority = static_cast<float>(i);
       auto sound_def_buffer = fpl::CreateSoundDef(builder, id, priority);
       builder.Finish(sound_def_buffer);
